Remove location description value object
