+++ source/Energinet.DataHub.MeteringPoints.Application/MarketDocuments/ValidationRuleSet.cs
@@ -35,7 +35,6 @@ namespace Energinet.DataHub.MeteringPoints.Application.MarketDocuments
             RuleFor(request => request).SetValidator(new MeasureUnitTypeRule());
             RuleFor(request => request).SetValidator(new MeterReadingOccurenceRule());
             RuleFor(request => request).SetValidator(new CapacityRule());
-            RuleFor(request => request).SetValidator(new LocationDescriptionMustBeValidRule());
             RuleFor(request => request).SetValidator(new PowerPlantMustBeValidRule());
             RuleFor(request => request).SetValidator(new ActualAddressRule());
             RuleFor(request => request).SetValidator(new AssetTypeRule());
