--- source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/LocationDescriptionMustBeValidRule.cs
@@ -1,36 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Energinet.DataHub.MeteringPoints.Application.Create;
-using Energinet.DataHub.MeteringPoints.Application.MarketDocuments;
-using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints.Rules;
-using FluentValidation;
-
-namespace Energinet.DataHub.MeteringPoints.Application.Validation.Rules
-{
-    public class LocationDescriptionMustBeValidRule : AbstractValidator<MasterDataDocument>
-    {
-        public LocationDescriptionMustBeValidRule()
-        {
-            When(request => !string.IsNullOrWhiteSpace(request.LocationDescription), () =>
-            {
-                RuleFor(createMeteringPoint => createMeteringPoint.LocationDescription)
-                    .Must(value => LocationDescription.CheckRules(value!).Success)
-                    .WithState(createMeteringPoint => new InvalidLocationDescriptionRuleError(createMeteringPoint.LocationDescription!));
-            });
-        }
-    }
-}
