+++ source/Energinet.DataHub.MeteringPoints.Tests/Domain/MeteringPoints/Exchange/CreationTests.cs
@@ -35,7 +35,6 @@ namespace Energinet.DataHub.MeteringPoints.Tests.Domain.MeteringPoints.Exchange
             var meteringPointGsrn = GsrnNumber.Create(SampleData.GsrnNumber);
             var meteringMethod = MeteringMethod.Virtual;
             var areadLinkId = new GridAreaLinkId(Guid.Parse(SampleData.GridAreaLinkId));
-            var locationDescription = LocationDescription.Create(string.Empty);
             var readingOccurrence = ReadingOccurrence.Hourly;
             var powerLimit = PowerLimit.Create(0, 0);
             var effectiveDate = EffectiveDate.Create(SampleData.EffectiveDate);
